+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - [new] Improve file location dialog ui
 - [new] Add an experimental option to run the app in the menu bar on macos
 - [new] Add an experimental option to auto start mac app (macos 13+ only)
+- [new] Auto select files and urls from pasteboard
+- [fix] Remove receive status after url received
 
 ### v2.0.146 19 November 2022
 - [new] Enable sending of multiple files